Save huff bits in huf0_0.huf in binary!!!!!!
1. fbitios.cpp is big change, I changed  it to make sure it is the same
as the code in the book;
2. fbitios.cpp can output one bit or read one bit in a time, which is
quite useful. I thought it can only operate date in BYTE level, but I am
wrong, it can operate date in BIT level. the key code in fbitios.cpp is
"if (outs.pos==7) ...."
because it is used to check if one byte is totally used, and the
function "put" in class fstream can only save byte each time. e.g.
fstream ofs, ofs.put(c).
3. I wrote two functions called Huffman_Encoding and Huffman_Decoding,
which will use RUN3D r[64] as a unit every time. These two functions
will feed the Hcodec one block (8*8) each time. the result of each block
will be saved in "xxx.huf" files in binary.
4. Problem now is that, in Escape encoding, the sign bit cannot be
saved. the first huffman encoded byte is "11000000"_2, where the first
seven bits are "1100000". But the start bit should be another '0' or
'1' (e.g. 01100000, the first bit 0 means the escaped encoded level in
RLE is positive.)
At the same time the second huffman escape encode byte is "00000000", I
am not sure if it is right. Because for escape encoding the encoded bits
should be like "01100000   000000", first 8 bits are explained and
following 6 bits are zeros runs in RLE.
5. I used a img (64*64) for test, which has 64 blocks (8*8).
most blocks can be decoded correctly after huffman encoding. but some of
them will be changed a lot which affect the visual quality seriously.
6. A big progress has been made today, Now is 4.44 AM on 7th,
April,2017. I have to go to sleep now, good night.